Allow for the timer to be invalidated so that we can pause. Stop works but we need to fix the pause so that it remembers the last paused time than rather continue (looks like time is skipping).

+++ Transcribr/Transcribr/TSViewController.m
@@ -12,6 +12,7 @@
 @interface TSViewController ()
 {
     TSAudioRecorder *audioRecorder;
+    NSTimer *timer;
 }
 @end
 
@@ -27,6 +28,7 @@
     [[self stopButton] setEnabled:NO];
     [[self pauseButton] setAlpha:GRAYED_OUT];
     [[self stopButton] setAlpha:GRAYED_OUT];
+    [self setPause:NO];
     
 }
 
@@ -46,8 +48,8 @@
     audioRecorder = [[TSAudioRecorder alloc] initWithFilename:dateString];
 //    [audioRecorder startRecording];
     
-    // create the timer, hold a reference to the timer if we want to cancel ahead of countdown
-    [NSTimer scheduledTimerWithTimeInterval:0.9 target:self selector:@selector(countdownUpdate:) userInfo:nil repeats:YES];
+    // Create timer, hold a reference to timer if we want to cancel ahead of countdown
+    timer = [NSTimer scheduledTimerWithTimeInterval:0.9 target:self selector:@selector(countdownUpdate:) userInfo:nil repeats:YES];
     
     // Ungray out pause & stop buttons and gray out record button
     [[self pauseButton] setEnabled:YES];
@@ -69,6 +71,8 @@
         // Ungray out pause & gray out rest
         [[self stopButton] setEnabled:NO];
         [[self stopButton] setAlpha:GRAYED_OUT];
+        [self setPause:YES];
+        [timer invalidate]; // Cancels the scheduledTimerWithTimeInterval
     }
     else {
         [[self pauseButton] setTitle:@"Pause" forState:UIControlStateNormal];
@@ -76,6 +80,9 @@
         // Ungray out pause & gray out rest
         [[self stopButton] setEnabled:YES];
         [[self stopButton] setAlpha:1];
+        
+        [self setPause:NO];
+        timer = [NSTimer scheduledTimerWithTimeInterval:0.9 target:self selector:@selector(countdownUpdate:) userInfo:nil repeats:YES]; // create timer, hold ref to timer if we want to cancel ahead of countdown
     }
 }
 
@@ -91,34 +98,39 @@
     if ([[[[self pauseButton] titleLabel] text] isEqualToString:@"Resume"]) {
         [[self pauseButton] setTitle:@"Pause" forState:UIControlStateNormal];
     }
+    
+    [timer invalidate]; // Cancels the scheduledTimerWithTimeInterval
+    [[self timeLeftLabel] setText:@"30:00"];
 }
 
 -(void)countdownUpdate:(NSTimer*) timer {
-    // code is written so one can see everything that is happening
-    // I am sure, some people would combine a few of the lines together
-    NSDate *currentDate = [NSDate date];
-    NSTimeInterval elaspedTime = [currentDate timeIntervalSinceDate:[self recordDate]];
-    
-    NSTimeInterval difference = MAX_RECORDING_TIME - elaspedTime;
-    if (difference <= 0) {
-        [timer invalidate];  // kill the timer
-        difference = 0;      // set to zero just in case iOS fired the timer late
-    }
-    int minutes = (int)difference/60;
-    int seconds =  (int)difference%60;
-    
-    // update the label with the remainding seconds
-    [[self timeLeftLabel] setText:[NSString stringWithFormat:@"%02d:%02d", minutes, seconds]];
-    
-    // Re-initialise the settings back to original if max recording time is reached
-    if (difference <= 0) {
-        [[self pauseButton] setEnabled:NO];
-        [[self stopButton] setEnabled:NO];
-        [[self recordButton] setEnabled:YES];
-        [[self pauseButton] setAlpha:GRAYED_OUT];
-        [[self stopButton] setAlpha:GRAYED_OUT];
-        [[self recordButton] setAlpha:1];
-        [[self timeLeftLabel] setText:@"30:00"];
+    if (![self pause]) {
+        // code is written so one can see everything that is happening
+        // I am sure, some people would combine a few of the lines together
+        NSDate *currentDate = [NSDate date];
+        NSTimeInterval elaspedTime = [currentDate timeIntervalSinceDate:[self recordDate]];
+        
+        NSTimeInterval difference = MAX_RECORDING_TIME - elaspedTime;
+        if (difference <= 0) {
+            [timer invalidate];  // kill the timer
+            difference = 0;      // set to zero just in case iOS fired the timer late
+        }
+        int minutes = (int)difference/60;
+        int seconds =  (int)difference%60;
+        
+        // update the label with the remainding seconds
+        [[self timeLeftLabel] setText:[NSString stringWithFormat:@"%02d:%02d", minutes, seconds]];
+        
+        // Re-initialise the settings back to original if max recording time is reached
+        if (difference <= 0) {
+            [[self pauseButton] setEnabled:NO];
+            [[self stopButton] setEnabled:NO];
+            [[self recordButton] setEnabled:YES];
+            [[self pauseButton] setAlpha:GRAYED_OUT];
+            [[self stopButton] setAlpha:GRAYED_OUT];
+            [[self recordButton] setAlpha:1];
+            [[self timeLeftLabel] setText:@"30:00"];
+        }
     }
 }
